# Dovio.Backend/openapi/mobile.v1.yaml
openapi: 3.1.0
info:
  title: Mobile API
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
# The rest of the spec mirrors the previous, without the old name
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /messages:
    get:
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: otherUserId
          schema: { type: string }
      responses: { '200': { description: OK } }
    post:
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiverId, content]
              properties:
                receiverId: { type: string }
                content: { type: string }
                mediaUrl: { type: string, format: uri }
      responses: { '201': { description: Created } }
  /messages/{messageId}:
    delete:
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: messageId
          required: true
          schema: { type: string }
      responses: { '200': { description: OK } }
  /messages/conversations:
    get:
      security: [ { bearerAuth: [] } ]
      responses: { '200': { description: OK } }
...